requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
black==25.12.0
certifi==2025.11.12
click==8.3.1
fastapi==0.125.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
mypy_extensions==1.1.0
numpy==2.4.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
ruff==0.14.10
scikit-learn==1.8.0
scipy==1.16.3
starlette==0.50.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
python-json-logger
prometheus-client
python-jose[cryptography]

